_archive/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
python-decouple==3.4
psycopg2-binary==2.8.6
django-rest-knox==4.1.0
gunicorn
django-heroku
django-storages
boto3
